+++ packages/validator/jest.config.js
@@ -8,13 +8,10 @@ module.exports = {
     displayName: name,
     name: name,
     rootDir: "../..",
+    projects: undefined,
     testMatch: [ `<rootDir>/packages/${dir}/src/**/*.test.js` ],
     collectCoverageFrom: [
         `<rootDir>/packages/${dir}/src/**/*.js`,
         `!<rootDir>/packages/${dir}/src/**/*.test.js`
-    ],
-    setupFiles: [
-        ...base.setupFiles,
-        `<rootDir>/packages/${dir}/jest.setup.js`
     ]
 };