+++ packages/utility/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/dlueth/qoopido/compare/@qoopido/utility@1.3.0...@qoopido/utility@1.3.1) (2020-06-22)
+
+**Note:** Version bump only for package @qoopido/utility
+
+
+
+
+
 # [1.3.0](https://github.com/dlueth/qoopido/compare/@qoopido/utility@1.2.0...@qoopido/utility@1.3.0) (2020-06-19)